refactor: refactor auth methods in a separate file
