+++ src/contexts/authContext.tsx
@@ -1,11 +1,8 @@
-import { createContext, useContext, useEffect, useReducer, useState } from "react";
+import { createContext, useContext, useEffect, useReducer } from "react";
 import { auth } from "../firebase.config";
-import { createUserWithEmailAndPassword, onAuthStateChanged } from "firebase/auth";
-import { AuthContextType, ReactChildrenType, ServicesType } from "../types";
+import { onAuthStateChanged } from "firebase/auth";
+import { AuthContextType, ReactChildrenType } from "../types";
 import { authReducer } from "../reducers/authReducer";
-import { signInService, signOutService, signUpService } from "../services/authService";
-import { useNavigate } from "react-router-dom";
-import toast from "react-hot-toast";
 
 const initialAuthState = {
       error: false,
@@ -15,8 +12,6 @@ const initialAuthState = {
             token: ""
       }
 };  
-
-const navigate = useNavigate();
 
 const AuthContext = createContext({} as AuthContextType);
 
@@ -33,51 +28,6 @@ export const AuthProvider = ({children}: ReactChildrenType) => {
                   }
             })
       }, [authDispatch])
-
-      const signUpHandler = async ( { email, password, firstName, lastName } : { email: string; password: string; firstName: string; lastName: string }) => {
-            authDispatch({type: "INITIALIZE"});
-            try {
-                  const response = await signUpService({email, password});
-                  const user = response.user;
-                  if(user) {
-                        navigate("/");
-                        toast.success("Account Created!");
-                        await createUser(user, { firstName, lastName });
-                  }
-            }
-            catch(err: any) {
-                  authDispatch({type: "SET_ERROR"});
-                  toast.error(err.message);
-            }
-      }
-      
-      const signInHandler = async ( {email, password} :ServicesType, from: { pathname: string }) => {
-            authDispatch({type: "INITIALIZE"});
-            try {
-                  const response = await signInService({email, password});
-                  const user = response.user;
-                  if(user) {
-                        navigate(from);
-                        toast.success("Login successfully!")
-                  }
-            }
-            catch(err: any) {
-                  authDispatch({type: "SET_ERROR"});
-                  toast.error(err.message);
-            }
-      }
-      
-      const signOutHandler = () => {
-            try {
-                  signOutService();
-                  navigate("/");
-                  toast.success("Signed out!");
-            }
-            catch(err: any) {
-                  authDispatch({type: "SET_ERROR"});
-                  toast.error(err.message);
-            }
-      }
 
       return (
             <AuthContext.Provider value={{ authState, authDispatch }}>
